+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added new parameters to the Physically Based Sky
 - Added Reflections to the DXR Wizard
 - Added a check in the custom post process template to throw an error if the default shader is not found.
+- Added the possibility to have ray traced colored and semi-transparent shadows on directional lights.
 
 ### Fixed
 - Sorting, undo, labels, layout in the Lighting Explorer.